only download if not already downloaded

+++ scripts/python/download_matches.py
@@ -1,4 +1,5 @@
 import csv
+import os
 from multiprocessing.pool import ThreadPool
 from time import time as timer
 
@@ -7,14 +8,15 @@ import requests
 
 def fetch_url(entry):
     filename, uri = entry
+    path = "data/2018/matches/{0}".format(filename)
 
-    r = requests.get(uri, stream=True)
-    if r.status_code == 200:
-        path = "data/2018/matches/{0}".format(filename)
-        with open(path, 'wb') as f:
-            for chunk in r:
-                f.write(chunk)
-        return path
+    if not os.path.exists(path):
+        r = requests.get(uri, stream=True)
+        if r.status_code == 200:
+            with open(path, 'wb') as f:
+                for chunk in r:
+                    f.write(chunk)
+    return path
 
 
 def matches_uri(end_of_uri):
@@ -25,7 +27,7 @@ with open("data/2018/matches.csv", "r") as teams_file:
     reader = csv.reader(teams_file, delimiter=",")
     next(reader)
 
-    urls = [("{0}.html".format(row[9].split("/")[-2].replace("match=","")), matches_uri(row[9])) for row in reader]
+    urls = [("{0}.html".format(row[1].split("/")[-2].replace("match=","")), matches_uri(row[1])) for row in reader]
 
 start = timer()
 results = ThreadPool(8).imap_unordered(fetch_url, urls)
